Move dht.Server defs to a separate file
